[core] move authentication_test_utils into ray._private to fix macos tests
